requirement.txt
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.2
Automat==0.6.0
channels==1.1.5
constantly==15.1.0
cycler==0.10.0
daphne==1.3.0
Django==1.11.2
django-pandas==0.4.3
django-registration-redux==1.6
hyperlink==17.2.1
incremental==17.5.0
matplotlib==2.0.2
mod-wsgi==4.5.15
mod-wsgi-httpd==2.4.23.2
msgpack-python==0.4.8
numpy==1.13.0
olefile==0.44
pandas==0.20.2
Pillow==4.1.1
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.6.0
pytz==2017.2
redis==2.10.5
RPi.GPIO==0.6.3
six==1.10.0
Twisted==17.5.0
txaio==2.8.0
XBee==2.3.1
zope.interface==4.4.2
